requirements.txt
tensorflow==2.17.0
tf-keras==2.17.0

protobuf==4.25.8
numpy==1.26.4
pillow==10.3.0
matplotlib==3.9.0
scikit-learn==1.5.0
streamlit==1.31.1
